week6/requirements.txt
kafka-python
confluent_kafka
requests
avro
faust-streaming
fastavro